requirements.txt
boto3
botocore
pydantic
fastapi
uvicorn
openai
numpy
pillow
opencv-python
matplotlib
requests
python-dotenv
pandas
ultralytics>=8.3.0
torch>=2.1.0
torchvision>=0.16.0
groundingdino
git+https://github.com/facebookresearch/segment-anything.git
